requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
google-generativeai
python-dotenv
xlrd
openpyxl
jinja2
matplotlib
seaborn
scipy
